# gateway-service/docs/organization/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Organization Service API",
        "title": "Organization Service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v2/org",
    "paths": {
...
        "/deal/{dealId}/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách sản phẩm thương vụ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thương vụ"
                ],
                "summary": "Lấy danh sách sản phẩm thương vụ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID thương vụ",
                        "name": "dealId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "schema": {
                            "$ref": "#/definitions/organizationpb.GetDealProductsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Thêm sản phẩm thương vụ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thương vụ"
                ],
                "summary": "Thêm sản phẩm thương vụ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID thương vụ",
                        "name": "dealId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Danh sách ID sản phẩm",
                        "name": "productIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "schema": {
                            "$ref": "#/definitions/organizationpb.SaveDealProductResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/org/deals/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tìm kiếm deals với các điều kiện lọc theo organizationId, groupId và text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thương vụ"
                ],
                "summary": "Tìm kiếm deals chung",
                "parameters": [
                    {
                        "description": "Thông tin tìm kiếm deals",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organizationpb.DealSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "schema": {
                            "$ref": "#/definitions/organizationpb.GetDealsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "authpb.Color": {
            "type": "object",
            "properties": {
                "backgroundColor": {
                    "type": "string"
                },
                "colorKey": {
                    "type": "string"
                },
                "contentColor": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "hexCode": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "integer"
                }
            }
        },
        "authpb.Permission": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/authpb.Permission"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "module": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "permissionType": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "authpb.Role": {
            "type": "object",
            "properties": {
                "allowAssign": {
                    "type": "boolean"
                },
                "color": {
                    "$ref": "#/definitions/authpb.Color"
                },
                "colorId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "domainType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isDefault": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "integer"
                },
                "permissionIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/authpb.Permission"
                    }
                },
                "roleDescription": {
                    "type": "string"
                },
                "roleGroupId": {
                    "type": "integer"
                },
                "roleGroupName": {
                    "type": "string"
                },
                "roleKey": {
                    "type": "integer"
                },
                "roleName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "organizationpb.BankAccountItem": {
            "type": "object",
            "properties": {
                "accountName": {
                    "type": "string"
                },
                "accountNumber": {
                    "type": "string"
                },
                "bank": {
                    "$ref": "#/definitions/paymentpb.Bank"
                },
                "bankId": {
                    "type": "integer"
                },
                "bankName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "organizationpb.DealSearchRequest": {
            "type": "object",
            "properties": {
                "groupId": {
                    "type": "integer"
                },
                "organizationId": {
                    "type": "integer"
                },
                "ownerId": {
                    "type": "integer"
                },
                "ownerOf": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "organizationpb.DocumentItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "organizationpb.GetDealProductsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bdspropb.ProductAttachment"
                    }
                }
            }
        },
        "organizationpb.GetDealsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/organizationpb.GroupDeal"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "organizationpb.GroupDeal": {
            "type": "object",
            "properties": {
                "allowSharing": {
                    "type": "boolean"
                },
                "bankAccount": {
                    "$ref": "#/definitions/organizationpb.BankAccountItem"
                },
                "cancelReason": {
                    "type": "string"
                },
                "chargePerson": {
                    "$ref": "#/definitions/sharepb.ProfileItem"
                },
                "chargePersonId": {
                    "type": "integer"
                },
                "chargePersonName": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sharepb.ProfileItem"
                    }
                },
                "dealType": {
                    "type": "integer"
                },
                "dealTypeName": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/organizationpb.DocumentItem"
                    }
                },
                "fromDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "internalNote": {
                    "type": "string"
                },
                "isManual": {
                    "description": "Nhập tay (dành cho doanh thu)",
                    "type": "boolean"
                },
                "isUnilateral": {
                    "description": "Đơn phương (gỡ khỏi thương vụ)",
                    "type": "boolean"
                },
                "memberCanAddTransaction": {
                    "type": "boolean"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sharepb.ProfileItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "numMember": {
                    "type": "integer"
                },
                "onlyOwnerGetCommission": {
                    "type": "boolean"
                },
                "ownerId": {
                    "type": "integer"
                },
                "ownerType": {
                    "type": "integer"
                },
                "partners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sharepb.ProfileItem"
                    }
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bdspropb.ProductAttachment"
                    }
                },
                "profitAmount": {
                    "type": "number"
                },
                "role": {
                    "$ref": "#/definitions/authpb.Role"
                },
                "status": {
                    "type": "integer"
                },
                "statusName": {
                    "type": "string"
                },
                "targetProfit": {
                    "type": "number"
                },
                "toDate": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "organizationpb.SaveDealProductResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "paymentpb.Bank": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "bdspropb.ProductAttachment": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "number"
                },
                "assetId": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "depositeId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "rentStatus": {
                    "type": "integer"
                },
                "saleStatus": {
                    "type": "integer"
                },
                "sourceType": {
                    "type": "integer"
                },
                "sourceTypeName": {
                    "type": "string"
                }
            }
        },
        "sharepb.ProfileItem": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isMember": {
                    "type": "boolean"
                },
                "joinedAt": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "tickVerified": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}